refactor(plugins): relocate memory remember/recall tool definitions into the plugin dir
Move the memory tools' definitions from tools/memory/register.ts into
plugins/defaults/memory/tools.ts so their memory-internal imports become
plugin-internal when the memory feature relocates into the plugin. They remain
CORE, always-loaded, workspace-overridable tools registered via the host tool
manifest, so behavior is byte-identical.

A plugin-tool *contribution* (the original approach) would have broken
core/workspace-override precedence and the always-loaded baseline, since plugin
tools register before the workspace scan and cannot be overridden/stripped by
workspace tools. Model-facing surfaces stay host-registered-but-plugin-defined.

Part of memory-as-a-plugin Track 2 (tools inbound-coupling decouple).
