+++ assistant/src/__tests__/plugin-import-boundary-guard.test.ts
@@ -130,6 +130,8 @@ const BASELINE: Record<string, readonly string[]> = {
     "../../../daemon/pkb-context-tracker.js",
     "../../../daemon/pkb-reminder-builder.js",
     "../../../daemon/trust-context.js",
+    "../../../memory/context-search/agent-runner.js",
+    "../../../memory/context-search/types.js",
     "../../../memory/graph/bootstrap.js",
     "../../../memory/graph/consolidation.js",
     "../../../memory/graph/conversation-graph-memory.js",
@@ -138,6 +140,8 @@ const BASELINE: Record<string, readonly string[]> = {
     "../../../memory/graph/graph-search.js",
     "../../../memory/graph/narrative.js",
     "../../../memory/graph/pattern-scan.js",
+    "../../../memory/graph/tool-handlers.js",
+    "../../../memory/graph/tools.js",
     "../../../memory/job-handlers/backfill.js",
     "../../../memory/job-handlers/embedding.js",
     "../../../memory/job-handlers/index-maintenance.js",
@@ -152,7 +156,10 @@ const BASELINE: Record<string, readonly string[]> = {
     "../../../memory/v2/consolidation-job.js",
     "../../../memory/v2/static-context.js",
     "../../../memory/v2/sweep-job.js",
+    "../../../permissions/types.js",
     "../../../persistence/jobs-store.js",
+    "../../../runtime/capabilities.js",
+    "../../../tools/types.js",
     "../../../types.js",
     "../../../util/logger.js",
     "../../../util/platform.js",
